rename `coins.rs` to `coin.rs`
